+++ CHANGELOG.md
@@ -2,7 +2,6 @@
 
 This changelog track changes to the qoqo project starting at version v0.5.0
 
-
 ## 1.16.0
 
 ### Fixed in 1.16.0
@@ -11,7 +10,7 @@ This changelog track changes to the qoqo project starting at version v0.5.0
 
 ### Added in 1.16.0
 
-* Added `InvSGate`, `InvTGate`, `SXGate`, `InvSXGate` gates.
+* Added `InvSGate`, `InvTGate`, `SXGate`, `InvSXGate`, `TripleControlledPauliX`, `TripleControlledPauliZ`, `TripleControlledPhaseShift`, `ControlledSWAP`, `PhaseShiftedControlledControlledZ`, `PhaseShiftedControlledControlledPhase` gates.
 
 ## 1.15.2
 
@@ -39,7 +38,7 @@ This changelog track changes to the qoqo project starting at version v0.5.0
 
 ### Added in 1.15.0
 
-* Added `SqrtPaulY` and `InvSqrtPauliY` gates.
+* Added `SqrtPauliY` and `InvSqrtPauliY` gates.
 
 ## 1.14.0